Split tests, fix ports init
